requirements.txt
acme==2.8.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.3
aiohttp-cors==0.7.0
aiohttp-fast-url-dispatcher==0.3.0
aiohttp-zlib-ng==0.3.1
aiooui==0.1.9
aiosignal==1.4.0
anyio==4.9.0
astral==2.2
async-interrupt==1.1.1
async-timeout==5.0.1
atomicwrites-homeassistant==1.4.1
attrs==23.2.0
awesomeversion==24.2.0
bcrypt==4.1.2
bleak==1.0.1
bleak-retry-connector==4.0.0
bluetooth-adapters==2.0.0
bluetooth-auto-recovery==1.5.2
bluetooth-data-tools==1.28.2
boto3==1.39.4
botocore==1.39.4
btsocket==0.3.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
ciso8601==2.3.1
cryptography==42.0.5
dbus-fast==2.44.1
ecdsa==0.19.1
envs==1.4
frozenlist==1.7.0
h11==0.16.0
habluetooth==4.0.1
hass-nabucasa==0.78.0
home-assistant-bluetooth==1.12.0
homeassistant==2024.3.3
httpcore==1.0.9
httpx==0.27.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.3
jmespath==1.0.1
josepy==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lru-dict==1.3.0
MarkupSafe==3.0.2
multidict==6.6.3
ocpp==2.0.0
orjson==3.9.15
packaging==25.0
propcache==0.3.2
pyasn1==0.6.1
pycognito==2023.5.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyRFC3339==2.0.1
PyRIC==0.1.6.3
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.1
referencing==0.36.2
requests==2.31.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
snitun==0.36.2
text-unidecode==1.3
typing_extensions==4.14.1
uart-devices==0.1.1
ulid-transform==0.9.0
urllib3==1.26.20
usb-devices==0.4.5
voluptuous==0.13.1
voluptuous-serialize==2.6.0
websockets==12.0
yarl==1.9.4
zlib-ng==0.5.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-aiohttp==1.0.5
pytest-mock==3.12.0
pytest-cov==4.1.0
aioresponses==0.7.6
